+++ src/StateSmith.Cli/Program.cs
@@ -42,27 +42,31 @@ class Program
             AnsiConsole.Console.Cursor.Show();
         };
 
+        int resultCode;
+
         try
         {
-            ParseCommandsAndRun(args, _console);
+            resultCode = ParseCommandsAndRun(args, _console);
         }
         catch (Exception ex)
         {
             _console.WriteException(ex);
-            Environment.ExitCode = 1;
+            resultCode = 1;
         }
         finally
         {
             _console.WriteLine("");
             _console.WriteLine("");
         }
+
+        Environment.ExitCode = resultCode;
     }
 
-    private void ParseCommandsAndRun(string[] args, IAnsiConsole _console)
+    private int ParseCommandsAndRun(string[] args, IAnsiConsole _console)
     {
         var parserResult = _cliArgsParser.Parse(args);
 
-        parserResult.MapResult(
+        int resultCode = parserResult.MapResult(
             (RunOptions opts) =>
             {
                 PreRunNoArgError(_console);
@@ -84,15 +88,36 @@ class Program
             },
             errs =>
             {
-                PrintHelp(parserResult, _console);
+                if (errs.Count() == 1 && errs.IsVersion())
+                {
+                    _console.WriteLine(HeadingInfo.Default);
+                    return 0;
+                }
+
+                //if (errs.Count() == 1 && errs.IsHelp())   // this breaks `ss.cli run --help`
+                if (args.Length == 1 && args[0] == "--help")
+                {
+                    _console.WriteLine(HeadingInfo.Default);
+                    _console.WriteLine();
+                    _console.WriteLine(_cliArgsParser.GetUsage());
+                    return 0;
+                }
+
                 if (errs.Count() == 1 && errs.First().Tag == ErrorType.NoVerbSelectedError)
                 {
+                    _console.WriteLine(HeadingInfo.Default);
+                    _console.WriteLine();
+                    _console.WriteLine(_cliArgsParser.GetUsage());
                     TryCheckForUpdates();
                     return ProvideMenu();
                 }
+
+                _console.WriteLine(_cliArgsParser.GetErrorHelp(parserResult, errs));
                 return 1;
             }
         );
+
+        return resultCode;
     }
 
     private int ProvideMenu()
@@ -155,18 +180,5 @@ class Program
         bool printed = loader.Printed || updateChecker.Printed;
         if (printed)
             _console.WriteLine("\n");
-    }
-
-    private static void PrintHelp(ParserResult<object> parserResult, IAnsiConsole _console)
-    {
-        var helpText = HelpText.AutoBuild(parserResult, h =>
-        {
-            h.AutoHelp = false;
-            h.AutoVersion = false;
-            h.Copyright = "";
-            return HelpText.DefaultParsingErrorsHandler(parserResult, h);
-        }, e => e);
-
-        _console.WriteLine(helpText);
     }
 }